requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==5.0.1
Faker==35.2.0
psycopg==3.2.3
psycopg-binary==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
